LeagueStats/requirements.txt
arrow==0.17.0
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
datapipelines==1.0.7
decorator==4.4.2
Django==3.1.5
django-cassiopeia==2.1.0
django-pgcrypto-fields==2.6.0
django-prometheus==2.1.0
django-redis==4.12.1
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
idna==2.10
jsonschema==3.2.0
lz4==3.1.2
merakicommons==1.0.9
networkx==2.5
Pillow==8.1.0
prometheus-client==0.10.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
riotwatcher==3.1.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
